Make DragSource manage drag start and end handlers
